Add files via upload
driver program, mainly

+++ Article.cs
@@ -8,9 +8,9 @@ namespace Everything
         private static int idSeed = 1;
         public int id { get; set; }
         public string name { get; set; }
-        public decimal averagePrice { get; set; }
+        public double averagePrice { get; set; }
         public int amount { get; set; }
-        public Article(int i, string n, decimal p, int a) {
+        public Article(int i, string n, double p, int a) {
             this.id = idSeed++;
             this.name = n;
             this.averagePrice = p;
