+++ Transaction.cs
@@ -8,11 +8,11 @@ namespace Everything
         private static int idSeed = 1;
         public int id { get; set; }
         public int articleId { get; set; }
-        public decimal price { get; set; }
+        public double price { get; set; }
         public int amount { get; set; }
         public DateTime dateTime { get; set; }
 
-        public Transaction(int id, int articleId,decimal price, int amount, DateTime dateTime) { 
+        public Transaction(int id, int articleId, double price, int amount, DateTime dateTime) { 
             this.id = idSeed++;
             this.articleId = articleId;
             this.price = price;
